Scale object proportionally with initial distance
