+++ src/hub.js
@@ -95,6 +95,7 @@ import "./components/open-media-button";
 import "./components/tweet-media-button";
 import "./components/remix-avatar-button";
 import "./components/transform-object-button";
+import "./components/scale-button";
 import "./components/hover-menu";
 import "./components/disable-frustum-culling";
 import "./components/teleporter";
